LogTag.java moved to logs folder to avoid clashes with other libraries
