+++ library/src/main/java/com/android/mms/dom/smil/SmilRegionElementImpl.java
@@ -16,7 +16,7 @@
 
 package com.android.mms.dom.smil;
 
-import com.android.mms.LogTag;
+import com.android.mms.logs.LogTag;
 
 import org.w3c.dom.DOMException;
 import org.w3c.dom.smil.SMILDocument;
