+++ library/src/main/java/com/android/mms/layout/LayoutManager.java
@@ -16,7 +16,7 @@
 
 package com.android.mms.layout;
 
-import com.android.mms.LogTag;
+import com.android.mms.logs.LogTag;
 
 import android.content.Context;
 import android.content.res.Configuration;
