+++ library/src/main/java/com/android/mms/transaction/MmsSystemEventReceiver.java
@@ -26,7 +26,7 @@ import android.os.Build;
 import android.provider.Telephony.Mms;
 import com.klinker.android.logger.Log;
 
-import com.android.mms.LogTag;
+import com.android.mms.logs.LogTag;
 import com.klinker.android.send_message.Utils;
 
 /**
