+++ library/src/main/java/com/android/mms/transaction/ReadRecTransaction.java
@@ -23,7 +23,7 @@ import android.net.Uri;
 import android.provider.Telephony.Mms.Sent;
 import com.klinker.android.logger.Log;
 
-import com.android.mms.LogTag;
+import com.android.mms.logs.LogTag;
 import com.google.android.mms.MmsException;
 import com.google.android.mms.pdu_alt.EncodedStringValue;
 import com.google.android.mms.pdu_alt.PduComposer;
