+++ library/src/main/java/com/android/mms/transaction/TransactionSettings.java
@@ -26,7 +26,7 @@ import com.android.mms.MmsConfig;
 import com.klinker.android.send_message.Transaction;
 import com.klinker.android.send_message.Utils;
 
-import com.android.mms.LogTag;
+import com.android.mms.logs.LogTag;
 
 /**
  * Container of transaction settings. Instances of this class are contained
