+++ library/src/main/java/com/android/mms/util/DownloadManager.java
@@ -26,11 +26,12 @@ import android.os.Handler;
 import android.os.Looper;
 import android.preference.PreferenceManager;
 import android.provider.Telephony.Mms;
+
+import com.android.mms.logs.LogTag;
 import com.klinker.android.logger.Log;
 import android.widget.Toast;
 
 import com.android.internal.telephony.TelephonyProperties;
-import com.android.mms.LogTag;
 import com.android.mms.service_alt.SystemPropertiesProxy;
 import com.google.android.mms.MmsException;
 import com.google.android.mms.pdu_alt.EncodedStringValue;
