+++ library/src/main/java/com/android/mms/util/SendingProgressTokenManager.java
@@ -16,10 +16,9 @@
 
 package com.android.mms.util;
 
-import com.android.mms.LogTag;
-
 import java.util.HashMap;
 
+import com.android.mms.logs.LogTag;
 import com.klinker.android.logger.Log;
 
 public class SendingProgressTokenManager {
